+++ src/Service/MediaConvert/src/Enum/Eac3CodingMode.php
@@ -10,6 +10,7 @@ final class Eac3CodingMode
     public const CODING_MODE_1_0 = 'CODING_MODE_1_0';
     public const CODING_MODE_2_0 = 'CODING_MODE_2_0';
     public const CODING_MODE_3_2 = 'CODING_MODE_3_2';
+    public const CODING_MODE_AUTO = 'CODING_MODE_AUTO';
     public const UNKNOWN_TO_SDK = 'UNKNOWN_TO_SDK';
 
     /**
@@ -21,6 +22,7 @@ final class Eac3CodingMode
             self::CODING_MODE_1_0 => true,
             self::CODING_MODE_2_0 => true,
             self::CODING_MODE_3_2 => true,
+            self::CODING_MODE_AUTO => true,
         ][$value]);
     }
 }