+++ src/Service/MediaConvert/src/ValueObject/VorbisSettings.php
@@ -8,8 +8,9 @@ namespace AsyncAws\MediaConvert\ValueObject;
 final class VorbisSettings
 {
     /**
-     * Optional. Specify the number of channels in this output audio track. Choosing Mono on the console gives you 1 output
-     * channel; choosing Stereo gives you 2. In the API, valid values are 1 and 2. The default value is 2.
+     * Optional. Specify the number of channels in this output audio track. Choosing Follow input will use the number of
+     * channels found in the audio source; choosing Mono on the console gives you 1 output channel; choosing Stereo gives
+     * you 2. In the API, valid values are 0, 1, and 2. The default value is 2.
      *
      * @var int|null
      */